no creation if no cube name

+++ app/api/views.py
@@ -133,9 +133,10 @@ def add_cube():
 @login_required
 def confirm_cube():
     temp_dir = os.path.join(TEMP_OLAPY_DIR, TEMP_CUBE_NAME)
-    if isdir(temp_dir):
-        # todo temp to fix
-        copy_tree(temp_dir, os.path.join(TEMP_OLAPY_DIR, 'cubes', request.data))
-        shutil.rmtree(temp_dir)
-        return jsonify({'success': True}), 200, {'ContentType': 'application/json'}
-    return jsonify({'success': False}), 400, {'ContentType': 'application/json'}
+    if request.data and request.method == 'POST':
+        if isdir(temp_dir):
+            # todo temp to fix
+            copy_tree(temp_dir, os.path.join(TEMP_OLAPY_DIR, 'cubes', request.data))
+            shutil.rmtree(temp_dir)
+            return jsonify({'success': True}), 200, {'ContentType': 'application/json'}
+        return jsonify({'success': False}), 400, {'ContentType': 'application/json'}
